Moved well layer types to types-file
